# docs/api/api-spec.yaml
swagger: "2.0"
info:
  description: "Auction Website API"
  version: "1.0.0"
  title: "Auction Website API"
host: "localhost:5000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /item_meta/categories/{id}:
    get:
      tags:
      - "item_meta"
      summary: "Get category by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the category to get.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Category"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
    put:
      tags:
      - "item_meta"
      summary: "Get category by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the category to get.
        - in: "body"
          name: "body"
          description: "Created category object"
          required: true
          schema:
            $ref: "#/definitions/CategoryReq"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Category"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
    
    delete:
      tags:
      - "item_meta"
      summary: "Delete category by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the category to delete.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Category"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  /item_meta/categories:
    post:
      tags:
      - "item_meta"
      summary: "Create new category"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created category"
        required: true
        schema:
          $ref: "#/definitions/CategoryReq"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Category"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  
    get:
      tags:
      - "item_meta"
      summary: "Get categories"
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "page"
        type: integer
        description: "Page number"
        required: true
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/CategoriesList"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
            
            
    delete:
      tags:
      - "item_meta"
      summary: "Delete ALL categories"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GenericResponse"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
...
  /auctions/{id}:
    get:
      tags:
      - "auction"
      summary: "Get auction by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the auction to get.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Auction"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
    put:
      tags:
      - "auction"
      summary: "Edit auction by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the auction to edit.
        - in: "body"
          name: "body"
          description: "Created auction object"
          required: true
          schema:
            $ref: "#/definitions/AuctionReq"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Auction"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
    
    delete:
      tags:
      - "auction"
      summary: "Delete auction by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the auction to delete.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Auction"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  /auctions:
    post:
      tags:
      - "auction"
      summary: "Create new auction"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created auction"
        required: true
        schema:
          $ref: "#/definitions/AuctionReq"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Auction"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  
    get:
      tags:
      - "auction"
      summary: "Get auctions"
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "page"
        description: "Page number"
        required: true
        type: integer
      - in: "query"
        name: "categoryId"
        description: "CategoryId"
        required: false
        type: integer
      - in: "query"
        name: "subcategoryId"
        description: "SubCategoryId"
        required: false
        type: integer
      - in: "query"
        name: "attributeId"
        description: "AttributeId"
        required: false
        type: integer
      - in: "query"
        name: "attributeValue"
        description: "AttributeValue"
        required: false
        type: string
      - in: "query"
        name: "initialPrice"
        description: "Initial Price"
        required: false
        type: number
      - in: "query"
        name: "sellerId"
        description: "seller id"
        required: false
        type: integer
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/AuctionsList"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
            
            
    delete:
      tags:
      - "auction"
      summary: "Delete ALL auctions"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GenericResponse"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  /bids/{id}:
    get:
      tags:
      - "bid"
      summary: "Get bid by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the bid to get.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Bid"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  
    delete:
      tags:
      - "bid"
      summary: "Delete bid by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the bid to delete.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Bid"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  /bids:
    post:
      tags:
      - "bid"
      summary: "Create new bid"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created bid"
        required: true
        schema:
          $ref: "#/definitions/BidReq"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Bid"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  
    get:
      tags:
      - "bid"
      summary: "Get bids"
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "page"
        description: "page number"
        required: true
        type: integer
      - in: "query"
        name: "sellerId"
        description: "seller id"
        required: false
        type: integer
      - in: "query"
        name: "bidderId"
        description: "bidder id"
        required: false
        type: integer
      - in: "query"
        name: "auctionId"
        description: "auction id"
        required: false
        type: integer
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/BidsList"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
            
            
    delete:
      tags:
      - "bid"
      summary: "Delete ALL bids"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GenericResponse"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
...
definitions:
...
  CategoryReq:
    type: "object"
    properties:
      categoryName:
        type: "string"
  Category:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      categoryName:
        type: "string"
  CategoriesList:
    type: "array"
    items:
      $ref: "#/definitions/Category"
...
  ItemAttribute:
    type: "object"
    properties:
      attributeId:
        type: "integer"
        format: "int32"
      attributeName:
        type: "string"
      attributeValue:
        type: "string"
  ItemAttributesList:
    type: "array"
    items:
      $ref: "#/definitions/ItemAttribute"
...
  Item:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      categoryId:
        type: "integer"
        format: "int32"
      categoryName:
        type: "string"
      subcategoryId:
        type: "integer"
        format: "int32"
      subcategoryName:
        type: "string"
      name:
        type: "string"
      attributes:
        $ref: "#/definitions/ItemAttributesList"
...
  AuctionReq:
    type: "object"
    properties:
      itemId:
        type: "integer"
        format: "int32"
      sellerId:
        type: "integer"
        format: "int32"
      initialPrice:
        type: "number"
        format: "float"
      minPrice:
        type: "number"
        format: "float"
      minIncrement:
        type: "number"
        format: "float"
      closingTime:
        type: "string"
  Auction:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      itemId:
        type: "integer"
        format: "int32"
      sellerId:
        type: "integer"
        format: "int32"
      initialPrice:
        type: "number"
        format: "float"
      minPrice:
        type: "number"
        format: "float"
      minIncrement:
        type: "number"
        format: "float"
      status:
        type: "string"
      currentHighestBid:
        type: "number"
        format: "float"
      winningBidId:
        type: "integer"
        format: "int32"
      closingTime:
        type: "string"
      item:
        $ref: "#/definitions/Item"
  AuctionsList:
    type: "array"
    items:
      $ref: "#/definitions/Auction"
...
  BidReq:
    type: "object"
    properties:
      auctionId:
        type: "integer"
        format: "int32"
      bidderId:
        type: "integer"
        format: "int32"
      bidValue:
        type: "number"
        format: "float"
      bidActive:
        type: "boolean"
  Bid:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      auctionId:
        type: "integer"
        format: "int32"
      bidderId:
        type: "integer"
        format: "int32"
      bidValue:
        type: "number"
        format: "float"
      bidActive:
        type: "boolean"
  BidsList:
    type: "array"
    items:
      $ref: "#/definitions/Bid"
...
  GenericResponse:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
    required:
      - code
      - message
  # Schema for error response body
  Error:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
    required:
      - code
      - message